% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffl_cleaners.R
\name{buffl_factorize_sociodemos}
\alias{buffl_factorize_sociodemos}
\title{Convert sociodemo variables to factors}
\usage{
buffl_factorize_sociodemos(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
the data frame with the fixed variables in the right format
}
\description{
Some variables are present in each and every buffl survey. This function converts these into the right format.
}
