% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffl_confints.R
\name{confint_multinomial_ordinal_effect}
\alias{confint_multinomial_ordinal_effect}
\title{Compute effects for a multinomial/ordinal model with one predictor}
\usage{
confint_multinomial_ordinal_effect(fit)
}
\arguments{
\item{fit}{the fitted model}
}
\value{
effect estimates with confidence interval
}
\description{
Compute effects for a multinomial/ordinal model with one predictor
}
\keyword{internal}
