% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRuffl.R
\docType{package}
\name{tRuffl}
\alias{tRuffl}
\alias{tRuffl-package}
\title{tRuffl: Interface Between Buffl And CleverR}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{tRuffl functions}{

The tRuffl functions ...
}

